DO-NOT-MERGE: Add ubifs_layout config option
Making ubifs_layout a config option, so a build could select if it needs
ubifs_layout feature or not.

Add example code to demo how a fixed layout image is updated by putting
it in /run/initramfs and reboot BMC to get it updated.

Tested: Generate a tarball with Romulus bmc image and manifest,
        update it with REST API (image upload and activation).

Signed-off-by: Lei YU <mine260309@gmail.com>